Incomplete: split lexer to external file
