+++ src/lib/core.js
@@ -3,7 +3,7 @@
 ///=========================================================================
 
 import {getOrInitObject, mergeObjects, parseJSON, varargConstructor} from "./utils.js";
-
+import {tokenize} from "./lexer/lexer.js"
 
 /**
  * @type {HyperscriptObject}
@@ -49,784 +49,6 @@ class ElementCollection {
 			[Symbol.iterator]();
 	}
 }
-
-//====================================================================
-// Lexer
-//====================================================================
-
-/** @type LexerObject */
-var _lexer = (function () {
-	var OP_TABLE = {
-		"+": "PLUS",
-		"-": "MINUS",
-		"*": "MULTIPLY",
-		"/": "DIVIDE",
-		".": "PERIOD",
-		"..": "ELLIPSIS",
-		"\\": "BACKSLASH",
-		":": "COLON",
-		"%": "PERCENT",
-		"|": "PIPE",
-		"!": "EXCLAMATION",
-		"?": "QUESTION",
-		"#": "POUND",
-		"&": "AMPERSAND",
-		$: "DOLLAR",
-		";": "SEMI",
-		",": "COMMA",
-		"(": "L_PAREN",
-		")": "R_PAREN",
-		"<": "L_ANG",
-		">": "R_ANG",
-		"<=": "LTE_ANG",
-		">=": "GTE_ANG",
-		"==": "EQ",
-		"===": "EQQ",
-		"!=": "NEQ",
-		"!==": "NEQQ",
-		"{": "L_BRACE",
-		"}": "R_BRACE",
-		"[": "L_BRACKET",
-		"]": "R_BRACKET",
-		"=": "EQUALS",
-	};
-
-	/**
-	 * isValidCSSClassChar returns `true` if the provided character is valid in a CSS class.
-	 * @param {string} c
-	 * @returns boolean
-	 */
-	function isValidCSSClassChar(c) {
-		return isAlpha(c) || isNumeric(c) || c === "-" || c === "_" || c === ":";
-	}
-
-	/**
-	 * isValidCSSIDChar returns `true` if the provided character is valid in a CSS ID
-	 * @param {string} c
-	 * @returns boolean
-	 */
-	function isValidCSSIDChar(c) {
-		return isAlpha(c) || isNumeric(c) || c === "-" || c === "_" || c === ":";
-	}
-
-	/**
-	 * isWhitespace returns `true` if the provided character is whitespace.
-	 * @param {string} c
-	 * @returns boolean
-	 */
-	function isWhitespace(c) {
-		return c === " " || c === "\t" || isNewline(c);
-	}
-
-	/**
-	 * positionString returns a string representation of a Token's line and column details.
-	 * @param {Token} token
-	 * @returns string
-	 */
-	function positionString(token) {
-		return "[Line: " + token.line + ", Column: " + token.column + "]";
-	}
-
-	/**
-	 * isNewline returns `true` if the provided character is a carrage return or newline
-	 * @param {string} c
-	 * @returns boolean
-	 */
-	function isNewline(c) {
-		return c === "\r" || c === "\n";
-	}
-
-	/**
-	 * isNumeric returns `true` if the provided character is a number (0-9)
-	 * @param {string} c
-	 * @returns boolean
-	 */
-	function isNumeric(c) {
-		return c >= "0" && c <= "9";
-	}
-
-	/**
-	 * isAlpha returns `true` if the provided character is a letter in the alphabet
-	 * @param {string} c
-	 * @returns boolean
-	 */
-	function isAlpha(c) {
-		return (c >= "a" && c <= "z") || (c >= "A" && c <= "Z");
-	}
-
-	/**
-	 * @param {string} c
-	 * @param {boolean} [dollarIsOp]
-	 * @returns boolean
-	 */
-	function isIdentifierChar(c, dollarIsOp) {
-		return c === "_" || c === "$";
-	}
-
-	/**
-	 * @param {string} c
-	 * @returns boolean
-	 */
-	function isReservedChar(c) {
-		return c === "`" || c === "^";
-	}
-
-	/**
-	 * @param {Token[]} tokens
-	 * @param {Token[]} consumed
-	 * @param {string} source
-	 * @returns {TokensObject}
-	 */
-	function makeTokensObject(tokens, consumed, source) {
-		consumeWhitespace(); // consume initial whitespace
-
-		/** @type Token | null */
-		var _lastConsumed = null;
-
-		function consumeWhitespace() {
-			while (token(0, true).type === "WHITESPACE") {
-				consumed.push(tokens.shift());
-			}
-		}
-
-		/**
-		 * @param {TokensObject} tokens
-		 * @param {*} error
-		 */
-		function raiseError(tokens, error) {
-			_parser.raiseParseError(tokens, error);
-		}
-
-		/**
-		 * @param {string} value
-		 * @returns {Token}
-		 */
-		function requireOpToken(value) {
-			var token = matchOpToken(value);
-			if (token) {
-				return token;
-			} else {
-				raiseError(this, "Expected '" + value + "' but found '" + currentToken().value + "'");
-			}
-		}
-
-		/**
-		 * @param {string} op1
-		 * @param {string} [op2]
-		 * @param {string} [op3]
-		 * @returns {Token | void}
-		 */
-		function matchAnyOpToken(op1, op2, op3) {
-			for (var i = 0; i < arguments.length; i++) {
-				var opToken = arguments[i];
-				var match = matchOpToken(opToken);
-				if (match) {
-					return match;
-				}
-			}
-		}
-
-		/**
-		 * @param {string} op1
-		 * @param {string} [op2]
-		 * @param {string} [op3]
-		 * @returns {Token | void}
-		 */
-		function matchAnyToken(op1, op2, op3) {
-			for (var i = 0; i < arguments.length; i++) {
-				var opToken = arguments[i];
-				var match = matchToken(opToken);
-				if (match) {
-					return match;
-				}
-			}
-		}
-
-		/**
-		 * @param {string} value
-		 * @returns {Token | void}
-		 */
-		function matchOpToken(value) {
-			if (currentToken() && currentToken().op && currentToken().value === value) {
-				return consumeToken();
-			}
-		}
-
-		/**
-		 * @param {string} type1
-		 * @param {string} [type2]
-		 * @param {string} [type3]
-		 * @param {string} [type4]
-		 * @returns {Token}
-		 */
-		function requireTokenType(type1, type2, type3, type4) {
-			var token = matchTokenType(type1, type2, type3, type4);
-			if (token) {
-				return token;
-			} else {
-				raiseError(this, "Expected one of " + JSON.stringify([type1, type2, type3]));
-			}
-		}
-
-		/**
-		 * @param {string} type1
-		 * @param {string} [type2]
-		 * @param {string} [type3]
-		 * @param {string} [type4]
-		 * @returns {Token | void}
-		 */
-		function matchTokenType(type1, type2, type3, type4) {
-			if (
-				currentToken() &&
-				currentToken().type &&
-				[type1, type2, type3, type4].indexOf(currentToken().type) >= 0
-			) {
-				return consumeToken();
-			}
-		}
-
-		/**
-		 * @param {string} value
-		 * @param {string} [type]
-		 * @returns {Token}
-		 */
-		function requireToken(value, type) {
-			var token = matchToken(value, type);
-			if (token) {
-				return token;
-			} else {
-				raiseError(this, "Expected '" + value + "' but found '" + currentToken().value + "'");
-			}
-		}
-
-		/**
-		 * @param {string} value
-		 * @param {string} [type]
-		 * @returns {Token | void}
-		 */
-		function matchToken(value, type) {
-			if (follows.indexOf(value) !== -1) {
-				return; // disallowed token here
-			}
-			var type = type || "IDENTIFIER";
-			if (currentToken() && currentToken().value === value && currentToken().type === type) {
-				return consumeToken();
-			}
-		}
-
-		/**
-		 * @returns {Token}
-		 */
-		function consumeToken() {
-			var match = tokens.shift();
-			consumed.push(match);
-			_lastConsumed = match;
-			consumeWhitespace(); // consume any whitespace
-			return match;
-		}
-
-		/**
-		 * @param {string} value
-		 * @param {string} [type]
-		 * @returns {Token[]}
-		 */
-		function consumeUntil(value, type) {
-			/** @type Token[] */
-			var tokenList = [];
-			var currentToken = token(0, true);
-
-			while (
-				(type == null || currentToken.type !== type) &&
-				(value == null || currentToken.value !== value) &&
-				currentToken.type !== "EOF"
-			) {
-				var match = tokens.shift();
-				consumed.push(match);
-				tokenList.push(currentToken);
-				currentToken = token(0, true);
-			}
-			consumeWhitespace(); // consume any whitespace
-			return tokenList;
-		}
-
-		/**
-		 * @returns {string}
-		 */
-		function lastWhitespace() {
-			if (consumed[consumed.length - 1] && consumed[consumed.length - 1].type === "WHITESPACE") {
-				return consumed[consumed.length - 1].value;
-			} else {
-				return "";
-			}
-		}
-
-		function consumeUntilWhitespace() {
-			return consumeUntil(null, "WHITESPACE");
-		}
-
-		/**
-		 * @returns {boolean}
-		 */
-		function hasMore() {
-			return tokens.length > 0;
-		}
-
-		/**
-		 * @param {number} n
-		 * @param {boolean} [dontIgnoreWhitespace]
-		 * @returns {Token}
-		 */
-		function token(n, dontIgnoreWhitespace) {
-			var /**@type {Token}*/ token;
-			var i = 0;
-			do {
-				if (!dontIgnoreWhitespace) {
-					while (tokens[i] && tokens[i].type === "WHITESPACE") {
-						i++;
-					}
-				}
-				token = tokens[i];
-				n--;
-				i++;
-			} while (n > -1);
-			if (token) {
-				return token;
-			} else {
-				return {
-					type: "EOF",
-					value: "<<<EOF>>>",
-				};
-			}
-		}
-
-		/**
-		 * @returns {Token}
-		 */
-		function currentToken() {
-			return token(0);
-		}
-
-		/**
-		 * @returns {Token | null}
-		 */
-		function lastMatch() {
-			return _lastConsumed;
-		}
-
-		/**
-		 * @returns {string}
-		 */
-		function sourceFor() {
-			return source.substring(this.startToken.start, this.endToken.end);
-		}
-
-		/**
-		 * @returns {string}
-		 */
-		function lineFor() {
-			return source.split("\n")[this.startToken.line - 1];
-		}
-
-		var follows = [];
-
-		function pushFollow(str) {
-			follows.push(str);
-		}
-
-		function popFollow() {
-			follows.pop();
-		}
-
-		function clearFollows() {
-			var tmp = follows;
-			follows = [];
-			return tmp;
-		}
-
-		function restoreFollows(f) {
-			follows = f;
-		}
-
-		/** @type {TokensObject} */
-		return {
-			pushFollow: pushFollow,
-			popFollow: popFollow,
-			clearFollow: clearFollows,
-			restoreFollow: restoreFollows,
-			matchAnyToken: matchAnyToken,
-			matchAnyOpToken: matchAnyOpToken,
-			matchOpToken: matchOpToken,
-			requireOpToken: requireOpToken,
-			matchTokenType: matchTokenType,
-			requireTokenType: requireTokenType,
-			consumeToken: consumeToken,
-			matchToken: matchToken,
-			requireToken: requireToken,
-			list: tokens,
-			consumed: consumed,
-			source: source,
-			hasMore: hasMore,
-			currentToken: currentToken,
-			lastMatch: lastMatch,
-			token: token,
-			consumeUntil: consumeUntil,
-			consumeUntilWhitespace: consumeUntilWhitespace,
-			lastWhitespace: lastWhitespace,
-			sourceFor: sourceFor,
-			lineFor: lineFor,
-		};
-	}
-
-	/**
-	 * @param {Token[]} tokens
-	 * @returns {boolean}
-	 */
-	function isValidSingleQuoteStringStart(tokens) {
-		if (tokens.length > 0) {
-			var previousToken = tokens[tokens.length - 1];
-			if (
-				previousToken.type === "IDENTIFIER" ||
-				previousToken.type === "CLASS_REF" ||
-				previousToken.type === "ID_REF"
-			) {
-				return false;
-			}
-			if (previousToken.op && (previousToken.value === ">" || previousToken.value === ")")) {
-				return false;
-			}
-		}
-		return true;
-	}
-
-	/**
-	 * @param {string} string
-	 * @param {boolean} [template]
-	 * @returns {TokensObject}
-	 */
-	function tokenize(string, template) {
-		var tokens = /** @type {Token[]}*/ [];
-		var source = string;
-		var position = 0;
-		var column = 0;
-		var line = 1;
-		var lastToken = "<START>";
-		var templateBraceCount = 0;
-
-		function inTemplate() {
-			return template && templateBraceCount === 0;
-		}
-
-		while (position < source.length) {
-			if (currentChar() === "-" && nextChar() === "-" && (isWhitespace(charAfterThat()) || charAfterThat() === "")) {
-				consumeComment();
-			} else {
-				if (isWhitespace(currentChar())) {
-					tokens.push(consumeWhitespace());
-				} else if (
-					!possiblePrecedingSymbol() &&
-					currentChar() === "." &&
-					(isAlpha(nextChar()) || nextChar() === "{")
-				) {
-					tokens.push(consumeClassReference());
-				} else if (
-					!possiblePrecedingSymbol() &&
-					currentChar() === "#" &&
-					(isAlpha(nextChar()) || nextChar() === "{")
-				) {
-					tokens.push(consumeIdReference());
-				} else if (currentChar() === "[" && nextChar() === "@") {
-					tokens.push(consumeAttributeReference());
-				} else if (currentChar() === "@") {
-					tokens.push(consumeShortAttributeReference());
-				} else if (isAlpha(currentChar()) || (!inTemplate() && isIdentifierChar(currentChar()))) {
-					tokens.push(consumeIdentifier());
-				} else if (isNumeric(currentChar())) {
-					tokens.push(consumeNumber());
-				} else if (!inTemplate() && (currentChar() === '"' || currentChar() === "`")) {
-					tokens.push(consumeString());
-				} else if (!inTemplate() && currentChar() === "'") {
-					if (isValidSingleQuoteStringStart(tokens)) {
-						tokens.push(consumeString());
-					} else {
-						tokens.push(consumeOp());
-					}
-				} else if (OP_TABLE[currentChar()]) {
-					if (lastToken === "$" && currentChar() === "{") {
-						templateBraceCount++;
-					}
-					if (currentChar() === "}") {
-						templateBraceCount--;
-					}
-					tokens.push(consumeOp());
-				} else if (inTemplate() || isReservedChar(currentChar())) {
-					tokens.push(makeToken("RESERVED", consumeChar()));
-				} else {
-					if (position < source.length) {
-						throw Error("Unknown token: " + currentChar() + " ");
-					}
-				}
-			}
-		}
-
-		return makeTokensObject(tokens, [], source);
-
-		/**
-		 * @param {string} [type]
-		 * @param {string} [value]
-		 * @returns {Token}
-		 */
-		function makeOpToken(type, value) {
-			var token = makeToken(type, value);
-			token.op = true;
-			return token;
-		}
-
-		/**
-		 * @param {string} [type]
-		 * @param {string} [value]
-		 * @returns {Token}
-		 */
-		function makeToken(type, value) {
-			return {
-				type: type,
-				value: value,
-				start: position,
-				end: position + 1,
-				column: column,
-				line: line,
-			};
-		}
-
-		function consumeComment() {
-			while (currentChar() && !isNewline(currentChar())) {
-				consumeChar();
-			}
-			consumeChar();
-		}
-
-		/**
-		 * @returns Token
-		 */
-		function consumeClassReference() {
-			var classRef = makeToken("CLASS_REF");
-			var value = consumeChar();
-			if (currentChar() === "{") {
-				classRef.template = true;
-				value += consumeChar();
-				while (currentChar() && currentChar() !== "}") {
-					value += consumeChar();
-				}
-				if (currentChar() !== "}") {
-					throw Error("Unterminated class reference");
-				} else {
-					value += consumeChar(); // consume final curly
-				}
-			} else {
-				while (isValidCSSClassChar(currentChar())) {
-					value += consumeChar();
-				}
-			}
-			classRef.value = value;
-			classRef.end = position;
-			return classRef;
-		}
-
-		/**
-		 * @returns Token
-		 */
-		function consumeAttributeReference() {
-			var attributeRef = makeToken("ATTRIBUTE_REF");
-			var value = consumeChar();
-			while (position < source.length && currentChar() !== "]") {
-				value += consumeChar();
-			}
-			if (currentChar() === "]") {
-				value += consumeChar();
-			}
-			attributeRef.value = value;
-			attributeRef.end = position;
-			return attributeRef;
-		}
-
-		function consumeShortAttributeReference() {
-			var attributeRef = makeToken("ATTRIBUTE_REF");
-			var value = consumeChar();
-			while (isValidCSSIDChar(currentChar())) {
-				value += consumeChar();
-			}
-			attributeRef.value = value;
-			attributeRef.end = position;
-			return attributeRef;
-		}
-
-		/**
-		 * @returns Token
-		 */
-		function consumeIdReference() {
-			var idRef = makeToken("ID_REF");
-			var value = consumeChar();
-			if (currentChar() === "{") {
-				idRef.template = true;
-				value += consumeChar();
-				while (currentChar() && currentChar() !== "}") {
-					value += consumeChar();
-				}
-				if (currentChar() !== "}") {
-					throw Error("Unterminated id reference");
-				} else {
-					consumeChar(); // consume final quote
-				}
-			} else {
-				while (isValidCSSIDChar(currentChar())) {
-					value += consumeChar();
-				}
-			}
-			idRef.value = value;
-			idRef.end = position;
-			return idRef;
-		}
-
-		/**
-		 * @returns Token
-		 */
-		function consumeIdentifier() {
-			var identifier = makeToken("IDENTIFIER");
-			var value = consumeChar();
-			while (isAlpha(currentChar()) || isIdentifierChar(currentChar())) {
-				value += consumeChar();
-			}
-			identifier.value = value;
-			identifier.end = position;
-			return identifier;
-		}
-
-		/**
-		 * @returns Token
-		 */
-		function consumeNumber() {
-			var number = makeToken("NUMBER");
-			var value = consumeChar();
-			while (isNumeric(currentChar())) {
-				value += consumeChar();
-			}
-			if (currentChar() === "." && isNumeric(nextChar())) {
-				value += consumeChar();
-			}
-			while (isNumeric(currentChar())) {
-				value += consumeChar();
-			}
-			number.value = value;
-			number.end = position;
-			return number;
-		}
-
-		/**
-		 * @returns Token
-		 */
-		function consumeOp() {
-			var op = makeOpToken();
-			var value = consumeChar(); // consume leading char
-			while (currentChar() && OP_TABLE[value + currentChar()]) {
-				value += consumeChar();
-			}
-			op.type = OP_TABLE[value];
-			op.value = value;
-			op.end = position;
-			return op;
-		}
-
-		/**
-		 * @returns Token
-		 */
-		function consumeString() {
-			var string = makeToken("STRING");
-			var startChar = consumeChar(); // consume leading quote
-			var value = "";
-			while (currentChar() && currentChar() !== startChar) {
-				if (currentChar() === "\\") {
-					consumeChar(); // consume escape char and move on
-				}
-				value += consumeChar();
-			}
-			if (currentChar() !== startChar) {
-				throw Error("Unterminated string at " + positionString(string));
-			} else {
-				consumeChar(); // consume final quote
-			}
-			string.value = value;
-			string.end = position;
-			string.template = startChar === "`";
-			return string;
-		}
-
-		/**
-		 * @returns string
-		 */
-		function currentChar() {
-			return source.charAt(position);
-		}
-
-		/**
-		 * @returns string
-		 */
-		function nextChar() {
-			return source.charAt(position + 1);
-		}
-
-		function charAfterThat() {
-			return source.charAt(position + 2);
-		}
-
-		/**
-		 * @returns string
-		 */
-		function consumeChar() {
-			lastToken = currentChar();
-			position++;
-			column++;
-			return lastToken;
-		}
-
-		/**
-		 * @returns boolean
-		 */
-		function possiblePrecedingSymbol() {
-			return (
-				isAlpha(lastToken) ||
-				isNumeric(lastToken) ||
-				lastToken === ")" ||
-				lastToken === "\"" ||
-				lastToken === "'" ||
-				lastToken === "`" ||
-				lastToken === "}" ||
-				lastToken === "]"
-			);
-		}
-
-		/**
-		 * @returns Token
-		 */
-		function consumeWhitespace() {
-			var whitespace = makeToken("WHITESPACE");
-			var value = "";
-			while (currentChar() && isWhitespace(currentChar())) {
-				if (isNewline(currentChar())) {
-					column = 0;
-					line++;
-				}
-				value += consumeChar();
-			}
-			whitespace.value = value;
-			whitespace.end = position;
-			return whitespace;
-		}
-	}
-
-	return {
-		tokenize: tokenize,
-		makeTokensObject: makeTokensObject,
-	};
-})();
 
 //====================================================================
 // Parser
@@ -1654,7 +876,6 @@ var _runtime = (function () {
 		var ctx = {
 			meta: {
 				parser: _parser,
-				lexer: _lexer,
 				runtime: _runtime,
 				owner: owner,
 				feature: feature,
@@ -1719,7 +940,7 @@ var _runtime = (function () {
 	* @returns {GrammarElement}
 	*/
 	function parse(src) {
-		var tokens = _lexer.tokenize(src);
+		var tokens = tokenize(src);
 		if (_parser.commandStart(tokens.currentToken())) {
 			var commandList = _parser.requireElement("commandList", tokens);
 			var last = commandList;
@@ -1819,7 +1040,7 @@ var _runtime = (function () {
 				try {
 					internalData.initialized = true;
 					internalData.script = src;
-					var tokens = _lexer.tokenize(src);
+					var tokens = tokenize(src);
 					var hyperScript = _parser.parseHyperScript(tokens);
 					if (!hyperScript) return;
 					hyperScript.apply(target || elt, elt);
@@ -2180,7 +1401,7 @@ var _runtime = (function () {
 		/** @type {any[]} */
 		var args;
 		if (stringToken.template) {
-			var innerTokens = _lexer.tokenize(rawValue, true);
+			var innerTokens = tokenize(rawValue, true);
 			args = parser.parseStringTemplate(innerTokens);
 		} else {
 			args = [];
@@ -2248,7 +1469,7 @@ var _runtime = (function () {
 		// TODO - unify these two expression types
 		if (elementId.template) {
 			var templateValue = elementId.value.substr(2, elementId.value.length - 2);
-			var innerTokens = _lexer.tokenize(templateValue);
+			var innerTokens = tokenize(templateValue);
 			var innerExpression = parser.requireElement("expression", innerTokens);
 			return {
 				type: "idRefTemplate",
@@ -2283,7 +1504,7 @@ var _runtime = (function () {
 		// TODO - unify these two expression types
 		if (classRef.template) {
 			var templateValue = classRef.value.substr(2, classRef.value.length - 2);
-			var innerTokens = _lexer.tokenize(templateValue);
+			var innerTokens = tokenize(templateValue);
 			var innerExpression = parser.requireElement("expression", innerTokens);
 			return {
 				type: "classRefTemplate",
@@ -2350,7 +1571,7 @@ var _runtime = (function () {
 
 		if (queryValue.indexOf("$") >= 0) {
 			var template = true;
-			var innerTokens = _lexer.tokenize(queryValue, true);
+			var innerTokens = tokenize(queryValue, true);
 			var args = parser.parseStringTemplate(innerTokens);
 		}
 
@@ -5374,7 +4595,6 @@ export default _hyperscript = mergeObjects(
 	},
 	{
 		internals: {
-			lexer: _lexer,
 			parser: _parser,
 			runtime: _runtime,
 		},
